wifi: rtl8xxxu: Fix RX skb size for aggregation disabled
commit from 15 July 2025 torvalds repo

Adjust RX skb size to include space for phystats increased the skb size
when aggregation is enabled but decreased it for the aggregation disabled case.

As a result, if a frame near the maximum size is received,
rtl8xxxu_rx_complete is called with status -EOVERFLOW and then the
driver starts to malfunction and no further communication is possible.

Restore the skb size in the aggregation disabled case.

Fixes: Adjust RX skb size to include space for phystats
Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Reviewed-by: Ping-Ke Shih <pkshih@realtek.com>
Signed-off-by: Kalicode Yogyakarta <kalicodejogja@mozmail.com>